Update to 4.5.0

+++ ChatSDKCore/Classes/Entities/PThread_.h
@@ -99,7 +99,7 @@ typedef enum {
 -(NSArray *) allMessages;
 -(BOOL) hasMessages;
 
--(NSArray *) loadMoreMessages: (int) numberOfMessages;
+-(NSArray *) loadMoreMessages: (NSInteger) numberOfMessages;
 -(void) resetMessages;
 
 -(NSDictionary *) metaDictionary;
